+++ src/shared/components/global-bottom-bar.tsx
@@ -77,7 +77,7 @@ interface ClientErrorLog {
 
 export function GlobalBottomBar({
   dbPermissionLevel,
-  siteVersion = "v2.1.703",
+  siteVersion = "v2.1.707",
 }: {
   dbPermissionLevel?: number;
   siteVersion?: string;